Extract NotificationsTable
